+++ src/app/inobeta-ui/ui/forms/array/array.ts
@@ -1,0 +1,35 @@
+import { IbFormControlBase } from "../controls/form-control-base";
+
+export interface IFormArrayOptions {
+  max?: number;
+}
+
+export interface IFormArray {
+  key: string;
+  fields: IbFormControlBase<any>[];
+  options?: IFormArrayOptions;
+  cols?: number;
+  rows?: number;
+  width?: string;
+}
+
+export class IbFormArray implements IFormArray {
+  key: string;
+  options?: IFormArrayOptions;
+  fields: IbFormControlBase<any>[];
+  cols: number;
+  rows: number;
+  width: string;
+  readonly role = "array";
+
+  constructor(options: IFormArray) {
+    this.key = options.key;
+    this.fields = options.fields ?? [];
+    this.options = options.options ?? {
+      max: Infinity,
+    };
+    this.cols = options.cols ?? 1;
+    this.rows = options.rows ?? 1;
+    this.width = options.width || '100%'
+  }
+}