+++ src/app/inobeta-ui/ui/forms/controls/form-control-base.ts
@@ -20,6 +20,7 @@ export class IbFormControlBase<T> {
   public cols: number;
   public rows: number;
   public debounceOnChange: number;
+  readonly role = 'control';
 
   constructor(options: IbFormControlBaseParams<T> = {}) {