+++ src/app/inobeta-ui/ui/forms/dynamic-form-array/dynamic-form-array.component.ts
@@ -1,0 +1,39 @@
+import { Component, Input } from "@angular/core";
+import { UntypedFormArray, UntypedFormGroup } from "@angular/forms";
+import { IbFormArray } from "../array/array";
+import { IbFormControlService } from "../form-control.service";
+
+@Component({
+  selector: "ib-dynamic-form-array",
+  templateUrl: "dynamic-form-array.component.html",
+})
+export class IbDynamicFormArrayComponent {
+  @Input() base: IbFormArray;
+  @Input() form: UntypedFormGroup;
+
+  constructor(private cs: IbFormControlService) {}
+
+  get array() {
+    return this.form.get(this.base.key) as UntypedFormArray;
+  }
+
+  get max() {
+    return this.base.options.max;
+  }
+
+  get lines() {
+    return this.array.controls.length;
+  }
+
+  getGroupAt(index: number) {
+    return this.array.at(index) as UntypedFormGroup;
+  }
+
+  addField() {
+    this.array.push(this.cs.toFormGroup(this.base.fields));
+  }
+
+  removeField(index: number) {
+    this.array.removeAt(index);
+  }
+}