+++ src/app/inobeta-ui/ui/forms/forms.module.ts
@@ -1,17 +1,24 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { ReactiveFormsModule } from '@angular/forms';
-import { IbDynamicFormControlComponent } from './dynamic-form-control/dynamic-form-control.component';
-import { IbDynamicFormComponent } from './dynamic-form/dynamic-form.component';
-import { IbFormControlService } from './form-control.service';
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { ReactiveFormsModule } from "@angular/forms";
+import { IbDynamicFormArrayComponent } from "./dynamic-form-array/dynamic-form-array.component";
+import { IbDynamicFormControlComponent } from "./dynamic-form-control/dynamic-form-control.component";
+import { IbDynamicFormComponent } from "./dynamic-form/dynamic-form.component";
+import { IbFormControlService } from "./form-control.service";
+import { IbFormPipeModule } from "./forms.pipes";
 
 @NgModule({
-  declarations: [IbDynamicFormControlComponent, IbDynamicFormComponent],
-  imports: [
-    CommonModule,
-    ReactiveFormsModule,
+  declarations: [
+    IbDynamicFormControlComponent,
+    IbDynamicFormComponent,
+    IbDynamicFormArrayComponent,
   ],
+  imports: [CommonModule, ReactiveFormsModule, IbFormPipeModule],
   providers: [IbFormControlService],
-  exports: [IbDynamicFormComponent, IbDynamicFormControlComponent]
+  exports: [
+    IbDynamicFormComponent,
+    IbDynamicFormControlComponent,
+    IbDynamicFormArrayComponent,
+  ],
 })
-export class IbDynamicFormsModule { }
+export class IbDynamicFormsModule {}