+++ src/app/inobeta-ui/ui/material-forms/material-form-array/material-form-array.component.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core";
+import { IbDynamicFormArrayComponent } from "../../forms/dynamic-form-array/dynamic-form-array.component";
+
+@Component({
+  selector: "ib-material-form-array",
+  templateUrl: "material-form-array.component.html",
+})
+export class IbMaterialFormArrayComponent extends IbDynamicFormArrayComponent {}