+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Inobeta/UI Changelog
 
+## [15.2.0] (2023-12-20)
+
+| Issue ID | Type | Section | Description |
+| -- | -- | -- | -- |
+|DEVK-223|feat|IbForm|Added support for FormArray|
+|DEVK-9|feat|IbHttp|Added directive and guard for role check|
+|DEVK-228|ci|Examples build|Fix examples build |
+
 ## [15.1.0] (2023-11-30)
 
 | Issue ID | Type | Section | Description |